+++ LCFrameworkDemo/TestPullLoadingViewController.m
@@ -176,7 +176,7 @@
     return 44;
 }
 
--(int) tableView:(LC_UITableView *)tableView numberOfRowsInSection:(NSInteger)section
+-(NSInteger) tableView:(LC_UITableView *)tableView numberOfRowsInSection:(NSInteger)section
 {
     return [LC_DS(___ds1) count];
 }
